requirements.txt
yfinance
ta
scikit-learn
pandas
requests
gspread
oauth2client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
